refactor: move TestDriftConfig to homeboy-extension-contract
Fifth sub-cut of the extension-contract campaign (Stage 1, wiki 12840).

TestDriftConfig is a fully pure serde struct (all fields Vec<String>/bool,
zero type dependencies) that lived inside manifest.rs. Moves it into a new
contract test_drift module; manifest.rs re-imports it so all six consumers
(manifest_config, test/drift, test/mod, defaults/builtins, +2) keep working
through the unchanged crate::extension::TestDriftConfig path.

This also partially unblocks a future manifest_config cut (TestDriftConfig was
one of its two blocking deps; the other, LifecycleContract, remains in core).
